+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master
 
+* `s3` - Support new `Aws::S3::EncryptionV2::Client` for client-side encryption (@janko)
+
 * `derivation_endpoint` – Avoid possibility of timing attacks when comparing signatures (@esparta)
 
 * `derivatives` – Avoid downloading the attached file when calling default no-op processor (@janko)